backend/requirements.txt
python-dotenv
pydantic
requests
google-api-python-client
tiktoken
mistralai>=0.0.11

fastapi
uvicorn
boto3

langgraph
langchain
langchain-google-genai
serpapi
google-search-results
litellm
google-generativeai
pinecone[grpc]
sentence_transformers
yfinance
snowflake-connector-python

selenium
webdriver-manager
pyarrow<19.0.0
snowflake-connector-python
matplotlib
seaborn
